Add auto-resizing of whisper dbs
Add a wrapper around contrib/whisper-auto-resize.py to check for the
need to resize our whisper databases -- basically if the configuration
has changed, call whisper-auto-resize.py before restarting carbon.

This naturally will slow the start-up of carbon in the cases where we
have made changes, but it does ensure that we never write to the wrong
format whisper file.