Add some tests to make sure grafana is ok-ish
I'd prefer there were more tests here, but wrangling with
a headless browser is a little more complex than i thought due to
Grafana being based on AngularJS.

For now, at least check that our files are deployed and that
nginx is started. This is equivalent to most of the other tests
in this formula.

More later, in other words.